figuring out why comments isnt working

+++ src/main/java/com/codeup/phaserun/repositories/RaceRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 
 @Repository
 public interface RaceRepository extends JpaRepository<Race, Long> {
-    Race findById(int id);
+    Race findById(long id);
     Race findRaceByRaceId(int id);
 }
